mcp-server/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary

langchain
langchain-core
langchain-openai
langchain-community
